feat: move file validation schemas to a separate file for better organization
